requirements.txt
altair==4.2.0
matplotlib==3.4.3
nltk==3.6.5
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.0
plotly==5.4.0
seaborn==0.11.2
streamlit==1.3.1
